[Claude] Add standalone .exe build with Nuitka + tree-sitter bundling
Build system for distributing Wiz as a single 36MB Windows executable.
Key challenge: tree-sitter .pyd bindings are Python C extensions that
Nuitka includes as data files but blocks via its import system. Fixed
by pre-loading into sys.modules via spec_from_file_location at startup.

- build_exe.py: Nuitka build script with selective binding inclusion
- config.py: is_bundled(), patch_tree_sitter_for_bundled()
- __main__.py: wiz init, scan timing, exe-mode severity, error messages
- UX polish across hooks, report, analyzer for bundled mode

+++ wiz/graph/depgraph.py
@@ -123,7 +123,7 @@ class DepGraph:
         result.extend(remaining)
         return result
 
-    def rank_by_importance(self) -> list[tuple[str, float]]:
+    def rank_by_importance(self) -> list[tuple[str, int]]:
         """Rank files by weighted fan_in/fan_out score.
 
         Score = fan_in * 2 + fan_out (being imported is more important).
@@ -133,7 +133,7 @@ class DepGraph:
             score = node.fan_in * 2 + node.fan_out
             ranked.append((path, score))
         ranked.sort(key=lambda x: (-x[1], x[0]))
-        return ranked  # type: ignore[return-value]
+        return ranked
 
     def to_dict(self) -> dict:
         return {
